fix: Made separate component for simulating staking
