+++ src/components/LiquidityMining/Create/Steps/PreviewAndCreate/index.tsx
@@ -1,4 +1,4 @@
-import React, { useCallback, useEffect, useMemo, useState } from 'react'
+import React, { useEffect, useState } from 'react'
 import { Box, Flex } from 'rebass'
 import {
   LiquidityMiningCampaign,
@@ -10,7 +10,7 @@ import {
 } from '@swapr/sdk'
 import PoolSummary from './PoolSummary'
 import RewardSummary from './RewardSummary'
-import { Card, Divider, SmoothGradientCard } from '../../../styleds'
+import { Card, Divider } from '../../../styleds'
 import { ButtonPrimary } from '../../../../Button'
 
 import styled from 'styled-components'
@@ -18,15 +18,8 @@ import { useActiveWeb3React } from '../../../../../hooks'
 import { CampaignCard } from '../../../../Pool/PairsList/CampaignCard'
 import { getStakedAmountUSD } from '../../../../../utils/liquidityMining'
 import { useNativeCurrencyUSDPrice } from '../../../../../hooks/useNativeCurrencyUSDPrice'
-import { TYPE } from '../../../../../theme'
-import { Repeat } from 'react-feather'
-import Slider from '../../../../Slider'
-import useDebouncedChangeHandler from '../../../../../utils/useDebouncedChangeHandler'
 
-import Loader from '../../../../Loader'
-import { parseUnits } from 'ethers/lib/utils'
-import { calculatePercentage } from '../../../../../utils'
-import { useTokenOrPairNativeCurrency } from '../../../../../hooks/useTokenOrPairNativeCurrency'
+import SimulateStaking from './SimulateStaking'
 
 const FlexContainer = styled(Flex)`
   ${props => props.theme.mediaWidth.upToExtraSmall`
@@ -46,37 +39,7 @@ const StyledCampaignCard = styled(CampaignCard)`
 const CampaignDetailWrapper = styled(Flex)`
   gap: 32px;
 `
-const SimulatedValue = styled.div`
-  font-family: 'Fira Mono';
-  font-style: normal;
-  font-weight: 700;
-  font-size: 18px;
-  line-height: 22px;
-  /* identical to box height */
 
-  text-align: right;
-  letter-spacing: 0.02em;
-  text-transform: uppercase;
-
-  color: ${props => props.theme.text2};
-`
-const StyledSwitch = styled(Repeat)`
-  width: 12px;
-  height: 12px;
-  stroke: ${props => props.theme.text5};
-  margin-left: 4px;
-`
-const SwitchContainer = styled.div`
-  font-size: 10px;
-  display: flex;
-  font-weight: 600;
-  color: ${props => props.theme.text5};
-  line-height: 11px;
-  letter-spacing: 0.15em;
-  text-transform: uppercase;
-  cursor: pointer;
-`
-const dollarAmountMaxSimulation = 10000000
 interface PreviewProps {
   campaign: SingleSidedLiquidityMiningCampaign | LiquidityMiningCampaign | null
   liquidityPair: Pair | Token | null
@@ -107,53 +70,7 @@ export default function PreviewAndCreate({
   const { account } = useActiveWeb3React()
   const [areButtonsDisabled, setAreButtonsDisabled] = useState(false)
   const { loading: loadingNativeCurrencyUsdPrice, nativeCurrencyUSDPrice } = useNativeCurrencyUSDPrice()
-  const [showUSDValue, setShowUSDValue] = useState(true)
 
-  const { loading: loadingNativeTokenPrice, derivedNativeCurrency: nativeTokenPrice } = useTokenOrPairNativeCurrency(
-    liquidityPair ? liquidityPair : undefined
-  )
-  const [simulatedValuePercentage, setSimulatedValuePercentage] = useState(0)
-
-  const maxStakedSimulatedAmount = useMemo(() => {
-    const base = stakingCap
-      ? parseFloat(stakingCap.multiply(nativeTokenPrice.multiply(nativeCurrencyUSDPrice)).toSignificant(22))
-      : dollarAmountMaxSimulation
-
-    const baseInUsd = parseFloat(nativeTokenPrice.multiply(nativeCurrencyUSDPrice).toFixed(22))
-
-    const baseValue = showUSDValue ? base : base / baseInUsd
-
-    const tokenOrPair = liquidityPair instanceof Token ? liquidityPair : liquidityPair?.liquidityToken
-
-    if (tokenOrPair && base !== 0 && baseInUsd !== 0) {
-      setSimulatedStakedAmount(
-        parseUnits(
-          calculatePercentage(base / baseInUsd, simulatedValuePercentage).toString(),
-          tokenOrPair.decimals
-        ).toString()
-      )
-    }
-
-    return calculatePercentage(baseValue, simulatedValuePercentage)
-  }, [
-    setSimulatedStakedAmount,
-    liquidityPair,
-    simulatedValuePercentage,
-    stakingCap,
-    nativeTokenPrice,
-    nativeCurrencyUSDPrice,
-    showUSDValue,
-  ])
-
-  const liquidityPercentChangeCallback = useCallback((value: number) => {
-    setSimulatedValuePercentage(value)
-  }, [])
-
-  const [innerLiquidityPercentage, setInnerLiquidityPercentage] = useDebouncedChangeHandler(
-    simulatedValuePercentage,
-    liquidityPercentChangeCallback,
-    10
-  )
   useEffect(() => {
     setAreButtonsDisabled(
       !!(!account || !reward || !liquidityPair || !startTime || !endTime || approvals || campaign === null)
@@ -168,9 +85,6 @@ export default function PreviewAndCreate({
   }
   const isSingleSided = campaign instanceof SingleSidedLiquidityMiningCampaign
 
-  const handleUSDValueClick = useCallback(() => {
-    setShowUSDValue(!showUSDValue)
-  }, [showUSDValue])
   return (
     <Flex flexDirection="column" style={{ zIndex: -1 }}>
       {campaign !== null && !loadingNativeCurrencyUsdPrice && (
@@ -190,35 +104,13 @@ export default function PreviewAndCreate({
             staked={true}
             campaign={campaign}
           />
-          <SmoothGradientCard
-            justifyContent={'space-between !important'}
-            flexDirection={'column'}
-            alignItems={'center'}
-            padding={'24px 28px'}
-            width={'50%'}
-          >
-            <TYPE.largeHeader fontSize={'11px'} letterSpacing="0.08em" color="text3">
-              SIMULATED STAKED AMOUNT
-            </TYPE.largeHeader>
-            <SwitchContainer onClick={handleUSDValueClick}>
-              Value in {showUSDValue ? 'crypto' : 'USD'}
-              <StyledSwitch />
-            </SwitchContainer>
-            {loadingNativeTokenPrice || loadingNativeCurrencyUsdPrice ? (
-              <Loader />
-            ) : (
-              <SimulatedValue>
-                {maxStakedSimulatedAmount.toLocaleString('en-us')}{' '}
-                {showUSDValue
-                  ? 'USD'
-                  : liquidityPair instanceof Token
-                  ? liquidityPair.symbol
-                  : `${liquidityPair?.token0.symbol}/${liquidityPair?.token1.symbol}`}
-              </SimulatedValue>
-            )}
-
-            <Slider value={innerLiquidityPercentage} size={16} onChange={setInnerLiquidityPercentage} />
-          </SmoothGradientCard>
+          <SimulateStaking
+            nativeCurrencyUSDPrice={nativeCurrencyUSDPrice}
+            loading={loadingNativeCurrencyUsdPrice}
+            setSimulatedStakedAmount={setSimulatedStakedAmount}
+            tokenOrPair={liquidityPair}
+            stakingCap={stakingCap}
+          />
         </CampaignDetailWrapper>
       )}
